+++ src/main/java/de/codescape/jira/plugins/scrumpoker/service/EstimationFieldService.java
@@ -1,18 +1,19 @@
 package de.codescape.jira.plugins.scrumpoker.service;
 
-import com.atlassian.jira.bc.issue.IssueService;
 import com.atlassian.jira.issue.CustomFieldManager;
-import com.atlassian.jira.issue.IssueInputParameters;
+import com.atlassian.jira.issue.IssueManager;
 import com.atlassian.jira.issue.MutableIssue;
+import com.atlassian.jira.issue.UpdateIssueRequest;
 import com.atlassian.jira.issue.fields.CustomField;
 import com.atlassian.jira.security.JiraAuthenticationContext;
 import com.atlassian.jira.user.ApplicationUser;
 import com.atlassian.plugin.spring.scanner.annotation.component.Scanned;
 import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import javax.inject.Named;
-import java.text.NumberFormat;
 
 /**
  * Component to access and write the value of the configured custom field. This is typically the Story Point field
@@ -21,6 +22,8 @@ import java.text.NumberFormat;
 @Named
 @Scanned
 public class EstimationFieldService {
+
+    private static final Logger log = LoggerFactory.getLogger(EstimationFieldService.class);
 
     private final ScrumPokerSettingsService scrumPokerSettingsService;
 
@@ -31,32 +34,32 @@ public class EstimationFieldService {
     private final JiraAuthenticationContext context;
 
     @ComponentImport
-    private final IssueService issueService;
+    private final IssueManager issueManager;
 
     @Inject
     public EstimationFieldService(JiraAuthenticationContext context,
                                   ScrumPokerSettingsService scrumPokerSettingsService,
-                                  IssueService issueService,
+                                  IssueManager issueManager,
                                   CustomFieldManager customFieldManager) {
         this.scrumPokerSettingsService = scrumPokerSettingsService;
+        this.issueManager = issueManager;
         this.customFieldManager = customFieldManager;
         this.context = context;
-        this.issueService = issueService;
     }
 
     /**
      * Save the estimation for a given issue.
      */
-    public void save(String issueKey, Integer newValue) {
+    public boolean save(String issueKey, Integer newValue) {
         ApplicationUser applicationUser = context.getLoggedInUser();
-        MutableIssue issue = issueService.getIssue(applicationUser, issueKey).getIssue();
-        IssueInputParameters issueInputParameters = issueService.newIssueInputParameters();
-        issueInputParameters.addCustomFieldValue(findStoryPointField().getIdAsLong(), formatAsNumber(newValue));
-
-        IssueService.UpdateValidationResult validationResult = issueService.validateUpdate(applicationUser,
-            issue.getId(), issueInputParameters);
-        if (!validationResult.getErrorCollection().hasAnyErrors()) {
-            issueService.update(applicationUser, validationResult);
+        MutableIssue issue = issueManager.getIssueByCurrentKey(issueKey);
+        issue.setCustomFieldValue(findStoryPointField(), newValue.doubleValue());
+        try {
+            issueManager.updateIssue(applicationUser, issue, UpdateIssueRequest.builder().build());
+            return true;
+        } catch (RuntimeException e) {
+            log.error("Could not save estimation {} for issue {}.", newValue, issueKey, e);
+            return false;
         }
     }
 
@@ -65,10 +68,6 @@ public class EstimationFieldService {
      */
     public CustomField findStoryPointField() {
         return customFieldManager.getCustomFieldObject(scrumPokerSettingsService.loadStoryPointFieldId());
-    }
-
-    private String formatAsNumber(Integer newValue) {
-        return NumberFormat.getInstance().format(newValue);
     }
 
 }