+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/EstimationFieldServiceTest.java
@@ -1,13 +1,12 @@
 package de.codescape.jira.plugins.scrumpoker.service;
 
-import com.atlassian.jira.bc.issue.IssueService;
 import com.atlassian.jira.issue.CustomFieldManager;
-import com.atlassian.jira.issue.IssueInputParameters;
+import com.atlassian.jira.issue.IssueManager;
 import com.atlassian.jira.issue.MutableIssue;
+import com.atlassian.jira.issue.UpdateIssueRequest;
 import com.atlassian.jira.issue.fields.CustomField;
 import com.atlassian.jira.security.JiraAuthenticationContext;
 import com.atlassian.jira.user.ApplicationUser;
-import com.atlassian.jira.util.ErrorCollection;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -16,8 +15,8 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-import java.util.HashMap;
-
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.*;
 
 public class EstimationFieldServiceTest {
@@ -34,7 +33,7 @@ public class EstimationFieldServiceTest {
     private JiraAuthenticationContext jiraAuthenticationContext;
 
     @Mock
-    private IssueService issueService;
+    private IssueManager issueManager;
 
     @Mock
     private CustomFieldManager customFieldManager;
@@ -49,66 +48,32 @@ public class EstimationFieldServiceTest {
     private ApplicationUser applicationUser;
 
     @Mock
-    private IssueService.IssueResult issueResult;
-
-    @Mock
-    private IssueInputParameters issueInputParameters;
-
-    @Mock
     private CustomField customField;
 
     @Mock
     private MutableIssue issue;
 
-    @Mock
-    private IssueService.UpdateValidationResult updateValidationResult;
-
-    @Mock
-    private ErrorCollection validationErrorCollection;
-
-    @Mock
-    private IssueService.IssueResult updateIssueResult;
-
-    @Mock
-    private ErrorCollection updateErrorCollection;
-
     @Before
     public void before() {
         when(jiraAuthenticationContext.getLoggedInUser()).thenReturn(applicationUser);
-        when(issueService.getIssue(applicationUser, ISSUE_KEY)).thenReturn(issueResult);
-        when(issueResult.getIssue()).thenReturn(issue);
+        when(issueManager.getIssueByCurrentKey(ISSUE_KEY)).thenReturn(issue);
         when(issue.getId()).thenReturn(ISSUE_ID);
-        when(issueService.newIssueInputParameters()).thenReturn(issueInputParameters);
         when(scrumPokerSettingsService.loadStoryPointFieldId()).thenReturn(CUSTOM_FIELD_ID);
         when(customFieldManager.getCustomFieldObject(CUSTOM_FIELD_ID)).thenReturn(customField);
-        when(issueService.validateUpdate(applicationUser, ISSUE_ID, issueInputParameters))
-            .thenReturn(updateValidationResult);
-        when(updateValidationResult.getErrorCollection()).thenReturn(validationErrorCollection);
     }
 
     @Test
-    public void withValidationErrorsIssueShouldNotBeUpdated() {
-        when(validationErrorCollection.hasAnyErrors()).thenReturn(true);
-        HashMap<String, String> errors = new HashMap<>();
-        errors.put("KEY-1", "Message 1");
-        errors.put("KEY-2", "Message 2");
-        when(validationErrorCollection.getErrors()).thenReturn(errors);
-
-        estimationFieldService.save(ISSUE_KEY, ESTIMATION);
-
-        verify(issueService, never()).update(applicationUser, updateValidationResult);
+    public void withErrorsDuringUpdateItShouldReturnFalse() {
+        when(issueManager.updateIssue(eq(applicationUser), eq(issue), any(UpdateIssueRequest.class)))
+            .thenThrow(new RuntimeException());
+        assertThat(estimationFieldService.save(ISSUE_KEY, ESTIMATION), is(false));
     }
 
     @Test
-    public void withoutValidationErrorsIssueShouldBeUpdated() {
-        when(validationErrorCollection.hasAnyErrors()).thenReturn(false);
-        when(issueService.update(applicationUser, updateValidationResult)).thenReturn(updateIssueResult);
-        when(updateIssueResult.getErrorCollection()).thenReturn(updateErrorCollection);
-        when(updateErrorCollection.hasAnyErrors()).thenReturn(false);
-
-        estimationFieldService.save(ISSUE_KEY, ESTIMATION);
-
-        verify(issueService, times(1)).update(applicationUser, updateValidationResult);
+    public void withSuccessfulUpdateItShouldReturnTrue() {
+        when(issueManager.updateIssue(eq(applicationUser), eq(issue), any(UpdateIssueRequest.class)))
+            .thenReturn(issue);
+        assertThat(estimationFieldService.save(ISSUE_KEY, ESTIMATION), is(true));
     }
 
 }